removing objects during eval like waymo

+++ 3DOpenWorldMOT/trainer.py
@@ -146,7 +146,7 @@ def load_model(cfg, checkpoints_dir, logger, rank=0):
                 else:
                     if rank == 0:
                         logger.info('No existing model, starting training from scratch...')
-
+            
             if cfg.training.optim.optimizer.o_class == 'Adam':
                 optimizer = torch.optim.Adam(
                     model.parameters(),
@@ -818,7 +818,8 @@ def eval_one_epoch(model, do_corr_clustering, rank, cfg, val_loader, experiment_
             else:
                 metric = detection_metric
 
-            # store weights if neural net                
+            # store weights if neural net 
+            print(metric[0], best_metric, metric[0] >= best_metric)
             if metric[0] >= best_metric:
                 best_metric = metric[0]
                 if is_neural_net and not cfg.just_eval:
@@ -877,9 +878,9 @@ def train(rank, cfg, world_size):
         load_model(cfg, checkpoints_dir, logger, rank)
     
     if rank == 0 or not cfg.multi_gpu:
-        if os.path.isdir(experiment_dir + name) and not cfg.continue_from_existing:
+        if not cfg.just_eval and os.path.isdir(experiment_dir + name) and not cfg.continue_from_existing:
             shutil.rmtree(experiment_dir + name)
-        if os.path.isdir(str(checkpoints_dir) + name) and not cfg.continue_from_existing:
+        if not cfg.just_eval and os.path.isdir(str(checkpoints_dir) + name) and not cfg.continue_from_existing:
             shutil.rmtree(str(checkpoints_dir) + name)
         os.makedirs(experiment_dir + name, exist_ok=True)
         logger.info(f'Detections are stored under {experiment_dir + name}...')
@@ -1044,7 +1045,7 @@ def train(rank, cfg, world_size):
    
     # final_evaluation
     dist.barrier()
-    if rank == 0:
+    if rank == 0 and not cfg.just_eval:
         for d in os.listdir(experiment_dir + name):
             shutil.rmtree(experiment_dir + name + f'/{d}')
     final_evaluation(
@@ -1069,9 +1070,9 @@ def train(rank, cfg, world_size):
 
 
 def on_return(rank, name, experiment_dir, checkpoints_dir, cfg):
-    if not cfg.keep_checkpoint and rank == 0:
+    if not cfg.just_eval and not cfg.keep_checkpoint and rank == 0:
         shutil.rmtree(checkpoints_dir + name)
-    if not cfg.keep_detections and rank == 0:
+    if not cfg.just_eval and not cfg.keep_detections and rank == 0:
         shutil.rmtree(experiment_dir + name)
